split up app.js, use modules
