+++ index.html
@@ -23,6 +23,8 @@
       <a href="#" role="button" id="sw-skip-waiting" class="alert-link">Update now</a> to get the latest version.
       <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
     </div>
+    <script src="js/updater.js" defer></script>
+
     <div id="app"></div>
     <p class="container">
       <a href="https://github.com/RyanBerliner/timesheets/blob/master/README.md" target="_blank">
@@ -35,6 +37,6 @@
     <script src="https://unpkg.com/react-dom@17.0.2/umd/react-dom.production.min.js" crossorigin></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js" crossorigin></script>
     <script src="https://cdn.jsdelivr.net/npm/bs5-react-elements@2.1.1/dist/bs5-react-elements.min.js" crossorigin></script>
-    <script src="app.js"></script>
+    <script src="js/app.js" type="module"></script>
   </body>
 </html>
