+++ service-worker.js
@@ -6,7 +6,9 @@ const ASSETS_STATIC = [
   'https://unpkg.com/react-dom@17.0.2/umd/react-dom.production.min.js',
   'https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js',
   'https://cdn.jsdelivr.net/npm/bs5-react-elements@2.1.1/dist/bs5-react-elements.min.js',
-  'app.js',
+  'js/app.js',
+  'js/state.js',
+  'js/updater.js',
   'index.html',
   'manifest.json',
   'icons-192.png',
